requirements.txt
langgraph==1.1.3
pdf2image==1.17.0
pymupdf==1.27.2.2
pypdf==6.9.2
pygraphviz==1.14
pyyaml==6.0.3
requests==2.33.0
openai>=1.0.0
pydantic>=2.0.0
pydantic-core
requests
pandas
numpy
scikit-learn
transformers
sentence-transformers
bert-score
cachetools
tqdm
pillow
pdfminer.six
tenacity
aiohttp
aioboto3
networkx
nano-vectordb
tiktoken
filelock
packaging
chromadb
pymilvus
pymongo
sqlalchemy
neo4j
oracledb
ollama
zai-sdk